Move hadSidebar focusing logic into useEditMode
